Debug 2, and no queue.
A relatively good debugger. The code is a lot better, and things seem to work quite well. We also removed the queue system, and count down twoards the fireing time instead of setting a constant fire time.

+++ Building_Brains_V4/include/usart_basic.h
@@ -36,7 +36,7 @@ extern "C" {
 #endif
 
 /* Normal Mode, Baud register value */
-#define USART0_BAUD_RATE(BAUD_RATE) ((float)(10000000.0 * 64 / (16 * (float)BAUD_RATE)) + 0.5)
+#define USART0_BAUD_RATE(BAUD_RATE) ((double)(10000000.0 * 64 / (16 * (double)BAUD_RATE)) + 0.5)
 
 int8_t USART_0_init();
